Rename metadata_class.py to metadata.py
